backend/requirements.txt
fastapi
uvicorn
python-dotenv
oracledb
requests
pydantic
loguru
redis
pytest
pydantic-settings